requirements.txt
flask==2.3.3
psycopg2-binary==2.9.7
requests==2.31.0
langchain-openai==0.0.2
langchain==0.0.350
langchain-community==0.0.10
faiss-cpu==1.7.4
python-dotenv==1.0.0
flask-cors==4.0.0
streamlit==1.28.1
pandas==2.1.1
openai==1.3.0 